+++ src/main/java/de/vette/idea/neos/lang/afx/AfxBundle.java
@@ -1,0 +1,28 @@
+
+package de.vette.idea.neos.lang.afx;
+
+import com.intellij.DynamicBundle;
+import org.jetbrains.annotations.Nls;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.PropertyKey;
+
+import java.util.function.Supplier;
+
+public final class AfxBundle extends DynamicBundle {
+    @NonNls
+    public static final String BUNDLE = "messages.AfxBundle";
+    private static final AfxBundle INSTANCE = new AfxBundle();
+
+    private AfxBundle() { super(BUNDLE); }
+
+    @NotNull
+    public static @Nls String message(@NotNull @PropertyKey(resourceBundle = BUNDLE) String key, Object @NotNull ... params) {
+        return INSTANCE.getMessage(key, params);
+    }
+
+    @NotNull
+    public static Supplier<@Nls String> messagePointer(@NotNull @PropertyKey(resourceBundle = BUNDLE) String key, Object @NotNull ... params) {
+        return INSTANCE.getLazyMessage(key, params);
+    }
+}