+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # IntelliJ Neos Changelog
 
 ## Unreleased
+- Add setting for AFX attribute value completion 
 
 ## 1.17.0 - 2023-06-07
 - Allow quotes in AFX attribute names